WIP - Windows support for build system.

+++ web/themes/contrib/civictheme/build.js
@@ -28,6 +28,7 @@ const config = {
   watch: false,
   cli: false,
   lintex: false,
+  install: false,
   combine: false,
   styles: false,
   styles_editor: false,
@@ -43,7 +44,7 @@ const config = {
 }
 
 const flags = process.argv.slice(2)
-if (['cli', 'lintex'].indexOf(flags[0]) >= 0) {
+if (['cli', 'lintex', 'install'].indexOf(flags[0]) >= 0) {
   config[flags[0]] = true
 } else {
   const buildType = ['build', 'watch']
@@ -74,6 +75,8 @@ let startTime = new Date().getTime()
 let lastTime = null
 
 const PATH = import.meta.dirname
+
+const ISWIN = process.platform === 'win32'
 
 const THEME_NAME                = PATH.split('/').reverse()[0]
 const DIR_CIVICTHEME            = fullPath('../../contrib/civictheme/')
@@ -116,10 +119,10 @@ const JS_ASSET_IMPORTS          = [
 const JS_LINT_EXCLUSION_HEADER  = '// phpcs:ignoreFile'
 
 const CONSTANTS_FILE_OUT        = `${DIR_OUT}/constants.json`
-const CONSTANTS_SCSS_IMPORTER   = fullPath(`./.storybook/importer.scss_variables.js`)
-const CONSTANTS_BACKGROUND_UTIL = `${COMPONENT_DIR}/00-base/background/background.utils.js`
-const CONSTANTS_ICON_UTIL       = `${COMPONENT_DIR}/00-base/icon/icon.utils.js`
-const CONSTANTS_LOGO_UTIL       = `${COMPONENT_DIR}/02-molecules/logo/logo.utils.js`
+const CONSTANTS_SCSS_IMPORTER   = importURL(fullPath(`./.storybook/importer.scss_variables.js`))
+const CONSTANTS_BACKGROUND_UTIL = importURL(`${COMPONENT_DIR}/00-base/background/background.utils.js`)
+const CONSTANTS_ICON_UTIL       = importURL(`${COMPONENT_DIR}/00-base/icon/icon.utils.js`)
+const CONSTANTS_LOGO_UTIL       = importURL(`${COMPONENT_DIR}/02-molecules/logo/logo.utils.js`)
 
 if (config.build) {
   build()
@@ -137,6 +140,10 @@ if (config.lintex) {
   lintExclusions()
 }
 
+if (config.install) {
+  install()
+}
+
 // ----------------------------------------------------------------------------- BUILD STEPS
 
 function buildOutDirectory() {
@@ -147,9 +154,15 @@ function buildOutDirectory() {
 
 function buildCombineDirectories() {
   if (config.combine && !config.base) {
-    runCommand(`rsync -a --delete ${DIR_UIKIT_COMPONENTS_IN}/ ${DIR_UIKIT_COPY_OUT}/`)
-    runCommand(`rsync -a --delete ${DIR_UIKIT_COPY_OUT}/ ${DIR_COMPONENTS_OUT}/`)
-    runCommand(`rsync -a ${DIR_COMPONENTS_IN}/ ${DIR_COMPONENTS_OUT}/`)
+    if (ISWIN) {
+      runCommand(`robocopy ${DIR_UIKIT_COMPONENTS_IN} ${DIR_UIKIT_COPY_OUT} /mir`)
+      runCommand(`robocopy ${DIR_UIKIT_COPY_OUT} ${DIR_COMPONENTS_OUT} /mir`)
+      runCommand(`robocopy ${DIR_COMPONENTS_IN} ${DIR_COMPONENTS_OUT} /mir`)
+    } else {
+      runCommand(`rsync -a --delete ${DIR_UIKIT_COMPONENTS_IN}/ ${DIR_UIKIT_COPY_OUT}/`)
+      runCommand(`rsync -a --delete ${DIR_UIKIT_COPY_OUT}/ ${DIR_COMPONENTS_OUT}/`)
+      runCommand(`rsync -a ${DIR_COMPONENTS_IN}/ ${DIR_COMPONENTS_OUT}/`)
+    }
     console.log(`Saved: Combined folders ${time()}`)
   }
 }
@@ -283,7 +296,11 @@ function buildJavascript() {
 
 function buildAssetsDirectory() {
   if (config.assets) {
-    runCommand(`rsync -a --delete --prune-empty-dirs --exclude .gitkeep --exclude js --exclude sass ${DIR_ASSETS_IN}/ ${DIR_ASSETS_OUT}/`)
+    if (ISWIN) {
+      runCommand(`robocopy ${DIR_ASSETS_IN} ${DIR_ASSETS_OUT} /s /purge /xf .gitkeep js sass`)
+    } else {
+      runCommand(`rsync -a --delete --prune-empty-dirs --exclude .gitkeep --exclude js --exclude sass ${DIR_ASSETS_IN}/ ${DIR_ASSETS_OUT}/`)
+    }
     console.log(`Saved: Assets ${time()}`)
   }
 }
@@ -365,26 +382,36 @@ function lintExclusions() {
   })
 }
 
+function install() {
+  if (ISWIN) {
+    runCommand(`robocopy .\\node_modules\\@civictheme\\uikit\\components\\ .\\components\\ /mir /unicode`)
+  } else {
+    runCommand(`rm -Rf components > /dev/null 2>&1 && cp -R node_modules/@civictheme/uikit/components components`)
+  }
+}
+
 // ----------------------------------------------------------------------------- UTILITIES
 
 function runCommand(command) {
-  execSync(command, (error, stdout, stderr) => {
-    if (error) {
-      console.log(`error: ${error.message}`)
-      return
+  try {
+    execSync(command)
+  } catch (err) {
+    if (command.indexOf('robocopy') === 0 && err.status < 2) {
+      // Do nothing. It was a success.
+      // https://learn.microsoft.com/en-us/windows-server/administration/windows-commands/robocopy#exit-return-codes
+    } else {
+      // Was likely an issue. Print output.
+      console.log(err.stdout.toString('utf8'))
     }
-    if (stderr) {
-      console.log(`stderr: ${stderr}`)
-      return
-    }
-    if (stdout) {
-      console.log(stdout)
-    }
-  })
+  }
 }
 
 function getImportsFromGlob(path, cwd) {
-  return globSync(path, { cwd }).sort((a, b) => a.localeCompare(b)).map(i => `@import "${i}";`).join('\n')
+  return globSync(path, { cwd })
+    .sort((a, b) => a.localeCompare(b))
+    .map(i => i.replaceAll('\\', '/'))
+    .map(i => `@import "${i}";`)
+    .join('\n')
 }
 
 function loadStyleFile(path, cwd) {
@@ -406,6 +433,10 @@ function stripJS(js) {
 
 function fullPath(filepath) {
   return path.resolve(PATH, filepath)
+}
+
+function importURL(url) {
+  return `${ISWIN ? 'file://' : ''}${ISWIN ? url.replaceAll('\\', '/') : url}`
 }
 
 function time(full) {
